refactor shop components directory
