--- src/components/shop/Post/index.tsx
@@ -1,4 +1,0 @@
-export * from './SaleTabPost'
-export * from './ReviewTabPost'
-export * from './LikeTabPost'
-export * from './OfferTabPost'
