--- src/components/shop/PostList/index.tsx
@@ -1,4 +1,0 @@
-export * from './OfferTabPostList'
-export * from './LikeTabPostList'
-export * from './SaleTabPostList'
-export * from './ReviewTabPostList'
