+++ src/components/shop/buy/OfferTabPost/index.tsx
