+++ src/components/shop/buy/OfferTabPost/styled.ts
