+++ src/components/shop/buy/OfferTabPost/types.ts
