+++ src/components/shop/buy/OfferTabPostList/index.tsx
@@ -2,7 +2,7 @@ import { Divider } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { Fragment } from 'react'
 import type { OfferPostListProps } from './types'
-import { OfferTabPost } from '@components/shop/Post'
+import { OfferTabPost } from '../OfferTabPost'
 
 const OfferTabPostList = ({
   className,
