+++ src/components/shop/buy/index.ts
@@ -1,0 +1,4 @@
+export * from './LikeTabPost'
+export * from './LikeTabPostList'
+export * from './OfferTabPost'
+export * from './OfferTabPostList'
