+++ src/components/shop/index.ts
@@ -1,8 +1,9 @@
-export * from './Post'
-export * from './PostList'
+export * from './sale'
+export * from './buy'
+export * from './review'
 export * from './ProfileBox'
 export * from './SelectBuyerModal'
 export * from './EditProfileModal'
-export * from './ReviewModal'
-export * from './ReviewModal/Write/types'
-export * from './ReviewModal/Read/types'
+export * from './buy/ReviewModal'
+export * from './buy/ReviewModal/Write/types'
+export * from './buy/ReviewModal/Read/types'
