+++ src/components/shop/review/index.ts
@@ -1,0 +1,2 @@
+export * from './ReviewTabPost'
+export * from './ReviewTabPostList'
