+++ src/components/shop/sale/SaleTabPost/types.ts
