+++ src/components/shop/sale/index.ts
@@ -1,0 +1,2 @@
+export * from './SaleTabPost'
+export * from './SaleTabPostList'
